refactor(ui-rewrite): move recent-activity mock out of the test tree
useRecentActivity (a runtime hook) imported RECENT_ACTIVITY_FIXTURE from
@/test/..., coupling app code to the test tree. Move it to src/mocks/ and
point the hook, the MSW handler, and the hook test at the new location. The
data stays gated behind VITE_USE_MOCK_ACTIVITY for the dev server.
